requirements.txt
aiofiles==24.1.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
Deprecated==1.2.18
get-annotations==0.1.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
nltk==3.9.1
packaging==25.0
protobuf==5.29.5
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
python-telegram-bot==22.1
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
wrapt==1.17.2
yandex-cloud-ml-sdk==0.11.0
yandexcloud==0.347.0
